requirements.txt
streamlit>=1.32.0
minio>=7.2.0
Pillow>=10.0.0
python-dotenv>=1.0.0
psycopg2-binary
pandas 